+++ Tests/VaporTests/SessionsTests.swift
@@ -3,11 +3,12 @@ import XCTest
 import Vapor
 import HTTP
 import Core
+import Cookies
 
 class SessionsTests: XCTestCase {
     static let allTests = [
         ("testExample", testExample),
-        ("testWithExpiryDate", testWithExpiryDate),
+        ("testCustomCookieFactoryWithExpiryDate", testCustomCookieFactoryWithExpiryDate),
     ]
 
     func testExample() throws {
@@ -61,9 +62,24 @@ class SessionsTests: XCTestCase {
         XCTAssertEqual(res2.body.bytes?.makeString(), "bar")
     }
     
-    func testWithExpiryDate() throws {
+    func testCustomCookieFactoryWithExpiryDate() throws {
         let s = MemorySessions()
-        let m = SessionsMiddleware(s)
+        let cookieName = "test-name"
+        let cookieFactory: (Request) -> Cookie = { req in
+            var cookie = Cookie(
+                name: cookieName,
+                value: "",
+                httpOnly: true
+            )
+            
+            if req.storage["session_expiry"] as? Bool ?? false {
+                let oneMonthTime: TimeInterval = 30 * 24 * 60 * 60
+                cookie.expires = Date().addingTimeInterval(oneMonthTime)
+            }
+            
+            return cookie
+        }
+        let m = SessionsMiddleware(s, cookieName: cookieName, cookieFactory: cookieFactory)
         let drop = try Droplet(middleware: [m])
         
         drop.get("should-set-expiry") { req in
@@ -75,7 +91,7 @@ class SessionsTests: XCTestCase {
         let req = Request(method: .get, path: "should-set-expiry")
         let res = try drop.respond(to: req)
         
-        guard let cookieIndex = res.cookies.index(of: "vapor-session") else {
+        guard let cookieIndex = res.cookies.index(of: cookieName) else {
             XCTFail("No cookie")
             return
         }
